requirements.txt
langchain==0.2.14
streamlit==1.37.1
openai==1.42.0
tiktoken==0.7.0
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
docx==0.2.4
python-docx==1.1.2